added parenthesis handling for each commands
