+++ srcs/parsing/commands/cmd_generator.c
@@ -6,72 +6,19 @@
 /*   By: oaarsse <oaarsse@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/09 15:30:26 by oaarsse           #+#    #+#             */
-/*   Updated: 2022/11/17 18:03:40 by oaarsse          ###   ########.fr       */
+/*   Updated: 2022/11/22 16:59:46 by oaarsse          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "parsing.h"
 
-static int	free_arg(char **arg)
-{
-	char	**tmp;
-
-	if (!arg)
-		return (-1);
-	tmp = arg;
-	while (*arg)
-	{
-		free(*arg);
-		arg++;
-	}
-	free(tmp);
-	return (-1);
-}
-
-static int	copy_args(char **src, char **dest)
-{
-	size_t	i;
-
-	if (!src || !dest)
-		return (-1);
-	i = 0;
-	while (*src)
-	{
-		*dest = ft_strdup(*src);
-		if (!*dest)
-			return (free_arg(dest));
-		i++;
-		dest++;
-		src++;
-	}
-	return (1);
-}
-
-static int	add_arg( t_command *command,
-		t_lst_tokens *tokens)
-{
-	size_t	len;
-	char	**new_arg;
-
-	len = 0;
-	while (command->args && command->args[len] != NULL)
-		len++;
-	new_arg = (char **)malloc(sizeof(char *) * (len + 2));
-	if (!new_arg)
-		return (-1);
-	new_arg[len] = ft_strdup(tokens->token);
-	if (!new_arg[len])
-		return (free_arg(new_arg));
-	new_arg[len + 1] = NULL;
-	if (!copy_args(command->args, new_arg))
-		return (free_arg(command->args));
-	free_arg(command->args);
-	command->args = new_arg;
-	return (1);
-}
-
 static int	initalize_cmd(t_command	*command, t_lst_tokens **tokens)
 {
+	if (is_separator((*tokens)->token) == e_OPEN_PRTH)
+	{
+		command->is_opening_parenthesis = 1;
+		*tokens = (*tokens)->next;
+	}
 	command->cmd = ft_strdup((*tokens)->token);
 	if (!command->cmd)
 		return (-1);
@@ -83,6 +30,21 @@ static int	initalize_cmd(t_command	*command, t_lst_tokens **tokens)
 		return (-1);
 	command->args[1] = NULL;
 	*tokens = (*tokens)->next;
+	return (1);
+}
+
+static int	final_touch(t_command	*command, t_lst_tokens **tokens,
+		t_token_separator sep)
+{
+	if (sep == e_OR || sep == e_AND || sep == e_PIPE || sep == e_CLOSE_PRTH)
+	{
+		if (sep == e_CLOSE_PRTH)
+			command->is_closing_parenthesis = 1;
+		else
+			command->e_sep = sep;
+		if (*tokens)
+			(*tokens) = (*tokens)->next;
+	}
 	return (1);
 }
 
@@ -101,16 +63,9 @@ int	command_generator(t_command	*command, t_lst_tokens **tokens)
 		else if (sep == e_NONE)
 			add_arg(command, *tokens);
 		else
-		{
 			if (handle_files(sep, command, tokens) == -1)
 				return (-1);
-		}
 		(*tokens) = (*tokens)->next;
 	}
-	if (sep == e_OR || sep == e_AND || sep == e_PIPE)
-	{
-		command->e_sep = sep;
-		(*tokens) = (*tokens)->next;
-	}
-	return (1);
+	return (final_touch(command, tokens, sep));
 }
